hopluccorp/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hopluccorp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Django
    "django==5.1.5",
    "django-environ==0.11.2",
    "django-model-utils==5.0.0",
    "django-allauth[socialaccount]==65.3.0",
    "django-crispy-forms==2.3",
    "crispy-bootstrap5==2024.10",
    "django-redis==5.4.0",
    "django-cors-headers==4.6.0",
    "django-filter==24.3",
    "django-celery-beat==2.7.0",
    "django-phonenumber-field[phonenumberslite]==8.0.0",

    # Django REST Framework
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "drf-spectacular==0.28.0",

    # Celery
    "celery==5.4.0",
    "flower==2.0.1",

    # Database
    "psycopg[binary]==3.2.4",

    # Cache & Message Broker
    "redis==5.2.1",

    # Auth
    "PyJWT==2.10.1",
    "cryptography==44.0.0",
    "social-auth-core==4.5.4",
    "social-auth-app-django==5.4.2",
    "google-auth==2.37.0",

    # Server
    "uvicorn[standard]==0.34.0",
    "gunicorn==23.0.0",

    # Utils
    "pillow==11.1.0",
    "python-slugify==8.0.4",
    "argon2-cffi==23.1.0",
    "whitenoise==6.8.2",
    "pydantic==2.10.5",
    "loguru==0.7.3",
    "geopy==2.4.1",
    "python-dateutil==2.9.0.post0",

    # CMS & Media
    "django-ckeditor-5==0.2.15",
    "cloudinary==1.41.0",
    "django-cloudinary-storage==0.3.0",
    "django-ordered-model==3.7.4",
    "django-cleanup==9.0.0",
    "django-modeltranslation==0.19.12",
]

[project.optional-dependencies]
dev = [
    "werkzeug[watchdog]>=3.1",
    "ipdb>=0.13",
    "watchfiles>=1.0",

    # Testing
    "pytest>=8.0",
    "pytest-django>=4.9",
    "pytest-sugar>=1.0",
    "factory-boy>=3.3",
    "coverage>=7.6",
    "django-coverage-plugin>=3.1",

    # Type checking
    "mypy>=1.12,<1.14",
    "django-stubs[compatible-mypy]>=5.1",
    "djangorestframework-stubs[compatible-mypy]>=3.15",

    # Code quality
    "ruff>=0.9",
    "pre-commit>=4.0",
    "djlint>=1.36",

    # Django dev tools
    "django-debug-toolbar>=4.4",
    "django-extensions>=3.2",
]

production = [
    "sentry-sdk>=2.19",
    "django-anymail[sendgrid]>=12.0",
]

# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db"
python_files = ["tests.py", "test_*.py"]

# ==== Coverage ====
[tool.coverage.run]
include = ["hopluccorp/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== Ruff (thay black + isort + flake8) ====
[tool.ruff]
target-version = "py312"
line-length = 119

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.ruff.lint.isort]
known-first-party = ["hopluccorp", "config"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"
